Use populate_obj to update objects

+++ admin/freeposte/admin/forms.py
@@ -11,8 +11,8 @@ class LoginForm(Form):
 
 class DomainForm(Form):
     name = fields.StringField('Domain name', [validators.DataRequired()])
-    max_users = fields_.DecimalField('Maximum user count', default=10)
-    max_aliases = fields_.DecimalField('Maximum alias count', default=10)
+    max_users = fields_.IntegerField('Maximum user count', default=10)
+    max_aliases = fields_.IntegerField('Maximum alias count', default=10)
     comment = fields.StringField('Comment')
     submit = fields.SubmitField('Create')
